More to change TimeTrigger to Trigger.
