+++ src/AutoHome.Data/SqliteDbContext.cs
@@ -11,5 +11,5 @@ public class SqliteDbContext : DbContext
 
     public DbSet<Device> Devices { get; set; } = null!;
 
-    public DbSet<Trigger> TimeTriggers { get; set; } = null!;
+    public DbSet<Trigger> Triggers { get; set; } = null!;
 }
