+++ src/AutoHome.Server/Endpoints/Triggers/Add.AddTriggerRequest.cs
@@ -9,5 +9,5 @@ public class AddTriggerRequest
     [Required]
     public string Name { get; set; } = string.Empty;
     [Required]
-    public TimeSpan Time { get; set; }
+    public double Interval { get; set; }
 }
