+++ src/AutoHome.Server/Endpoints/Triggers/List.ListTriggersResult.cs
@@ -5,5 +5,5 @@ public class ListTriggersResult
     public Guid Id { get; set; }
     public Guid DeviceId { get; set; }
     public string Name { get; set; } = string.Empty;
-    public TimeSpan Time { get; set; }
+    public double Interval { get; set; }
 }
