+++ src/AutoHome.Server/Endpoints/Triggers/Update.UpdateTriggerResult.cs
@@ -5,5 +5,5 @@ public class UpdateTriggerResult
     public Guid Id { get; set; }
     public Guid DeviceId { get; set; }
     public string Name { get; set; } = string.Empty;
-    public TimeSpan Time { get; set; }
+    public long Interval { get; set; }
 }
